feat: task, task list, day, toolbar styles refactor
