+++ README.md
@@ -1,1 +1,41 @@
-# React
+# Test task for Addax
+
+Task Calendar. It is user friendly to create tasks and move theme to any date in calendar.
+
+You are able to run application with : [Vercel deploy link](https://test-task-addax-calendar.vercel.app/)
+
+Or, you may want to run it locally:
+
+## Clone repository
+
+```bash
+git clone https://github.com/chertoha/test-task-addax-calendar.git
+```
+
+## Environment Variables
+
+Create an `.env` file in the root directory and add the following variables.
+This project is a frontend application built with React. It uses environment variables to connect to an API.
+
+```bash
+VITE_APP_HOLIDAYS_API_URL= //holidays API here
+
+```
+
+VITE_APP_HOLIDAYS_API_URLL Used to connect to API for holidays date requests
+
+## Run the development server:
+
+```bash
+npm run dev
+# or
+yarn dev
+# or
+pnpm dev
+# or
+bun dev
+```
+
+## Overview
+
+Open [http://localhost:5000](http://localhost:5000) with your browser to see the result.
