+++ src/components/AddTaskButton/AddTaskButton.tsx
@@ -1,42 +1,11 @@
-import { addTask } from "@/redux/tasks/slice";
-import { TaskType } from "@/types/entities";
-import { nanoid } from "@reduxjs/toolkit";
 import { FC } from "react";
 import { useDispatch } from "react-redux";
-import styled from "styled-components";
+import { nanoid } from "@reduxjs/toolkit";
 import { IoMdAdd } from "react-icons/io";
 
-export const ButtonWrapper = styled("div")`
-  width: 20px;
-  height: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-export const Button = styled("button")`
-  /* display: none; */
-
-  opacity: 0;
-  transform: translate(0, -30px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  width: 100%;
-
-  border-radius: 50%;
-  border: 2px solid #0f7d0d;
-
-  color: #0f7d0d;
-
-  & > svg {
-    height: 100%;
-    width: 100%;
-  }
-
-  transition: transform 400ms ease-in-out, opacity 300ms ease-in-out;
-`;
+import { addTask } from "@/redux/tasks/slice";
+import { TaskType } from "@/types/entities";
+import { Button, ButtonWrapper } from "./AddTaskButton.styled";
 
 interface IProps {
   date: Date;
