+++ src/components/TaskList/TaskList.styled.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const List = styled("ul")`
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+`;
